refactor: js 파일 모듈화
